Move unrelated train stuff away from ORM resource module
